pyproject.toml
[project]
name = "autopatcher"
version = "0.0.1"
description = "This project is for running Autopkg and importing into munki as well as promoting new versions through munki catalogs"
authors = [{ name = "AnywhereOps", email = "andrew@anywhereops.ai" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "arrow==1.4.0",
    "certifi==2025.11.12",
    "cloud-autopkg-runner==0.24.13",
    "gitpython==3.1.45",
    "packaging==25.0",
    "pygithub==2.8.1",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "slack-sdk==3.37.0",
]

[project.urls]
Homepage = "https://AnywhereOps.github.io/autopatcher/"
Repository = "https://github.com/AnywhereOps/autopatcher"
Documentation = "https://AnywhereOps.github.io/autopatcher/"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.23.0",
    "ty>=0.0.1a16",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true
include = ["scripts/**/*.py", "tests/**/*.py"]
extend-exclude = ["AutoPkg"]
force-exclude = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # StartProcessWithPartialPath - intentionally use PATH lookup for git
    "S607",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["scripts"]
